Clean up dissection of "power requested" and "power available" TLVs.
Fix the addition of power values to the top-level item for the TLV so
that it actually adds power values.

Make the list of power values in that item display correctly, without
extra commas.

Fail if the length of the TLV is less than 8.  (We should really add an
expert info item for that.)

Change-Id: Ic4229c0652306f69156b8341c9fbb67cacc8154c
Reviewed-on: https://code.wireshark.org/review/25215
Reviewed-by: Guy Harris <guy@alum.mit.edu>