Update the introductory comment.
Update the URL for the Cisco CMTS.

Remove the stuff about the Arris devices; the link doesn't work, their
site doesn't seem to mention "Cadant", and the C4 page I did find
doesn't say anything about how to do captures (that's probably hidden
behind a customerwall).

Emphasize that, if you can, and do, select "DOCSIS" as the link-layer
type when capturing on an Ethernet, the file (for pcap) or interface
(for pcapng) will be handled as DOCSIS, even if you don't set the "force
all frames to be dissected as DOCSIS" preference.

Change-Id: Iec44d4b9c6af231e9b7f36962515b566b3433fb4
Reviewed-on: https://code.wireshark.org/review/25777
Reviewed-by: Guy Harris <guy@alum.mit.edu>