Fix endianness.
Make the 1-byte fields little-endian (it doesn't matter, but it's more
consistent that way), and make the transmission attenuation fields
little-endian (where it *does* matter - making them big-endian was
probably a copy-and-pasteo).

Note that the OUI field being "big-endian" is correct, however.
(Radiotap specifies it as an array of 3 octets containing an OUI, and we
display OUIs as big-endian 24-bit quantities.)

Change-Id: I42d19f7ec0d066ce89dbef78d11dff900c0a6b60
Reviewed-on: https://code.wireshark.org/review/25998
Reviewed-by: Guy Harris <guy@alum.mit.edu>