Remove a bunch of obsolete code.
We no longer have the code to create a bundle, as we rely on CMake
having done so, at least to the extent of populating the bundle with all
the files we've generated.  Get rid of the code that used to support it,
and the command-line options that are no longer necessary now that we no
longer build code bundles.

Don't have explicit lists of CLI or extcap binaries; instead, just look
for all plain files in Wireshark.app/Contents/MacOS that have read and
execute permissions for owner/group/user.  That way, we don't have to
update the script if we add new binaries or new directories of binaries.

Change-Id: I047296a7889bea71165eebde10f34bec6ea96cc5
Reviewed-on: https://code.wireshark.org/review/26032
Reviewed-by: Guy Harris <guy@alum.mit.edu>