Just use tvb_memeql() to check the signature.
That's one of the things it's intended to do - check the values of parts
of the packet that might not actually be there.  The comparison fails if
either 1) it's all there and doesn't match or 2) it's not all there.

Change-Id: I0f97ea5f75c2cada511e254ec096d294ea710f45
Reviewed-on: https://code.wireshark.org/review/26040
Reviewed-by: Guy Harris <guy@alum.mit.edu>
(cherry picked from commit 017167fb34220a17af361d50b2d239927acac7c1)
Reviewed-on: https://code.wireshark.org/review/26041