Don't handle various "command to send" values in the default case.
The default case ignores the high-order bit, which is set in all the
values for "command to send", so they will never be matched.  The values
moved out of the default case, if their upper bit is clear, either don't
correspond to any command in T.30 or correspond to an initial
identification command, which never has the upper bit set, so there's no
risk of misidentification by processing all of the "command to send"
values outside the default case.

Thanks and a tip of the Hatlo hat to Visual Studio Code Analysis for
catching this one.

Change-Id: I6192b0c5a6dcfd31b9fd757be736a311a9d089e6
Reviewed-on: https://code.wireshark.org/review/26198
Reviewed-by: Guy Harris <guy@alum.mit.edu>