Eliminate unused structure members and an unused list.
We don't use the addresses, and don't use the inbound port.

We also don't need a list of tn5250_conv_info_t's - they're allocated
with wmem, so they get freed when we're done with a file, and they're
attached to conversations, so we don't need a list to find them.

Change-Id: Id3b44fb11cf353d7d15a58938883f2460cf642a0
Reviewed-on: https://code.wireshark.org/review/26301
Reviewed-by: Guy Harris <guy@alum.mit.edu>