Register and initialize the ZCL clusters in a uniform way.
The cluster registration and initialisation is done in the functions 'proto_reg_handoff_zbee_zcl_XXX'.
In these functions the 1st step was to register the dissector.
Then the cluster was initialised via the function 'zbee_zcl_init_cluster'.

In this patch the registration and initialisation is now done by the function  'zbee_zcl_init_cluster'.
In this way we have a more uniform way of registration.

Change-Id: If865d904ea51c299a48e5c0004f3f280b7479d49
Reviewed-on: https://code.wireshark.org/review/26325
Petri-Dish: Alexis La Goutte <alexis.lagoutte@gmail.com>
Tested-by: Petri Dish Buildbot
Reviewed-by: Alexis La Goutte <alexis.lagoutte@gmail.com>