Fix another leak, which happens with -E occurrence=l.
For each occurrence, if there was already an occurrence in the array, we
were just removing it. not freeing what it pointed to.

While we're at it, expand comments. and always check the array size with
"!= 0", not "> 0" - the value is unsigned, so they're equivalent, but
this makes the code more self-consistent.

Change-Id: I538f46b296a7721a39ba4366c2f6269e7e097b0d
Reviewed-on: https://code.wireshark.org/review/26328
Reviewed-by: Guy Harris <guy@alum.mit.edu>
(cherry picked from commit 9a5217bdd4680e36f2e92838348bd4ea010d28a0)
Reviewed-on: https://code.wireshark.org/review/26329