When looking up response codes, just keep the key on the stack.
We don't need to dynamically allocate a guint variable to hold the
response code value - and then not bother to free it! - if we're just
doing a lookup; we can use an automatic variable and pass a pointer to
it.

Change-Id: I6edbb352f0aa33c91ede0f4e1bbb962fa921bea2
Reviewed-on: https://code.wireshark.org/review/26375
Reviewed-by: Guy Harris <guy@alum.mit.edu>
(cherry picked from commit 7ffc4f96ebe2fac503cb251a848441ad8d4bad8c)
Reviewed-on: https://code.wireshark.org/review/26376