Add support for ZigBee ZCL manufacturer specific attributes and commands.
For each ZCL cluster there is the possibility to implement manufacturer specific attributes and commands.
This is done by setting a flag (Manufacturer specific = true/false) in the Frame Control Field of the ZCL frame.
If this flag is set then also the Manufacturer Code is added to the ZCL frame.
Example: Manufacturer Code = 0x100b = Philips.
So basically this means that the meaning of an attribute/command is dependent on the manufacturer code.

Currently the ZCL frame is parsed and the manufacturer code is also dissected and displayed correctly.
However, the dissection of manufacturer specific attributes and commands is not done properly.
Their is no distinction between a global attribute/command and a manufacturer specific attribute/command.
This results in the incorrect dissection of these manufacturer specific attributes/commands.

This patch fixes this issue.
The registration of a cluster is now done with a cluster Id AND the manufacturer specific code.

Change-Id: I5d4aeb5473fd56f2a570d2a1f83d1090d42423ab
Reviewed-on: https://code.wireshark.org/review/26440
Petri-Dish: Anders Broman <a.broman58@gmail.com>
Tested-by: Petri Dish Buildbot
Reviewed-by: Anders Broman <a.broman58@gmail.com>